https://github.com//issues/556 お知らせ情報登録処理部分をcomponent化
